+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :bug:`275` Fix acoustic transfer `one_source` bug to take dead volume into account.
 * :bug:`274` Container.wells_with(prop, val) not returning wells with val due to `is` operator instead of `==`
 
 * :release:`7.3.0 <2020-10-01>`